BuxFix - Moved to Bootstrap 4, Completed other issues from the review
