+++ pom.xml
@@ -86,7 +86,7 @@
 		<dependency>
 			<groupId>org.webjars</groupId>
 			<artifactId>bootstrap</artifactId>
-			<version>3.4.1</version>
+			<version>4.5.0</version>
 		</dependency>
