+++ src/main/java/com/jm/online_store/repository/RoleRepository.java
@@ -6,5 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.Optional;
 
 public interface RoleRepository extends JpaRepository<Role, Long> {
+
     Optional<Role> findByName(String name);
 }
